# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_UniverseHappiness_WiseDx_internal_errors.AppError:
    properties:
      code:
        $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_UniverseHappiness_WiseDx_internal_errors.ErrorCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 2000
    - 2001
    - 2002
    - 2003
    - 2004
    - 2100
    - 2101
    - 2102
    - 2103
    type: integer
    x-enum-varnames:
    - ErrBadRequest
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrMethodNotAllowed
    - ErrConflict
    - ErrTooManyRequests
    - ErrInternalServer
    - ErrServiceUnavailable
    - ErrTimeout
    - ErrValidation
    - ErrTenantNotFound
    - ErrTenantAlreadyExists
    - ErrTenantInactive
    - ErrTenantNameRequired
    - ErrTenantInvalidStatus
    - ErrAgentMissingThinkingModel
    - ErrAgentMissingAllowedTools
    - ErrAgentInvalidMaxIterations
    - ErrAgentInvalidTemperature
  github_com_UniverseHappiness_WiseDx_internal_types.AgentConfig:
    properties:
      allowed_tools:
        description: List of allowed tool names
        items:
          type: string
        type: array
      knowledge_bases:
        description: Accessible knowledge base IDs
        items:
          type: string
        type: array
      knowledge_ids:
        description: Accessible knowledge IDs (individual documents)
        items:
          type: string
        type: array
      max_iterations:
        description: Maximum number of ReAct iterations
        type: integer
      reflection_enabled:
        description: Whether to enable reflection
        type: boolean
      system_prompt_web_disabled:
        description: Custom prompt when web search is disabled
        type: string
      system_prompt_web_enabled:
        description: Custom prompt when web search is enabled
        type: string
      temperature:
        description: LLM temperature for agent
        type: number
      use_custom_system_prompt:
        description: Whether to use custom system prompt instead of default
        type: boolean
      web_search_enabled:
        description: Whether web search tool is enabled
        type: boolean
      web_search_max_results:
        description: 'Maximum number of web search results (default: 5)'
        type: integer
    type: object
...
  github_com_UniverseHappiness_WiseDx_internal_types.ContextCompressionStrategy:
    enum:
    - sliding_window
    - smart
    type: string
    x-enum-varnames:
    - ContextCompressionSlidingWindow
    - ContextCompressionSmart
  github_com_UniverseHappiness_WiseDx_internal_types.ContextConfig:
    properties:
      compression_strategy:
        allOf:
        - $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.ContextCompressionStrategy'
        description: 'Compression strategy: "sliding_window" or "smart"'
      max_tokens:
        description: Maximum tokens allowed in LLM context
        type: integer
      recent_message_count:
        description: |-
          For sliding_window: number of messages to keep
          For smart: number of recent messages to keep uncompressed
        type: integer
      summarize_threshold:
        description: 'Summarize threshold: number of messages before summarization'
        type: integer
    type: object
  github_com_UniverseHappiness_WiseDx_internal_types.ConversationConfig:
    properties:
      context_template:
        description: ContextTemplate is the prompt template for summarizing retrieval
          results
        type: string
      embedding_top_k:
        type: integer
      enable_query_expansion:
        type: boolean
      enable_rewrite:
        type: boolean
      fallback_prompt:
        type: string
      fallback_response:
        type: string
      fallback_strategy:
        description: Fallback strategy
        type: string
      keyword_threshold:
        type: number
      max_completion_tokens:
        description: MaxTokens is the maximum number of tokens to generate
        type: integer
      max_rounds:
        description: Retrieval & strategy parameters
        type: integer
      prompt:
        type: string
      rerank_model_id:
        type: string
      rerank_threshold:
        type: number
      rerank_top_k:
        type: integer
      rewrite_prompt_system:
        description: Rewrite prompts
        type: string
      rewrite_prompt_user:
        type: string
      summary_model_id:
        description: Model configuration
        type: string
      temperature:
        description: Temperature controls the randomness of the model output
        type: number
      use_custom_context_template:
        type: boolean
      use_custom_system_prompt:
        description: Prompt is the system prompt for normal mode
        type: boolean
      vector_threshold:
        type: number
    type: object
...
  github_com_UniverseHappiness_WiseDx_internal_types.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  github_com_UniverseHappiness_WiseDx_internal_types.LoginResponse:
    properties:
      message:
        type: string
      refresh_token:
        type: string
      success:
        type: boolean
      tenant:
        $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.Tenant'
      token:
        type: string
      user:
        $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.User'
    type: object
...
  github_com_UniverseHappiness_WiseDx_internal_types.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  github_com_UniverseHappiness_WiseDx_internal_types.RegisterResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
      tenant:
        $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.Tenant'
      user:
        $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.User'
    type: object
  github_com_UniverseHappiness_WiseDx_internal_types.RetrieverEngineParams:
    properties:
      retriever_engine_type:
        allOf:
        - $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.RetrieverEngineType'
        description: Retriever engine type
      retriever_type:
        allOf:
        - $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.RetrieverType'
        description: Retriever type
    type: object
  github_com_UniverseHappiness_WiseDx_internal_types.RetrieverEngineType:
    enum:
    - postgres
    - elasticsearch
    - infinity
    - elasticfaiss
    - qdrant
    type: string
    x-enum-varnames:
    - PostgresRetrieverEngineType
    - ElasticsearchRetrieverEngineType
    - InfinityRetrieverEngineType
    - ElasticFaissRetrieverEngineType
    - QdrantRetrieverEngineType
  github_com_UniverseHappiness_WiseDx_internal_types.RetrieverEngines:
    properties:
      engines:
        items:
          $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.RetrieverEngineParams'
        type: array
    type: object
...
  github_com_UniverseHappiness_WiseDx_internal_types.Tenant:
    properties:
      agent_config:
        allOf:
        - $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.AgentConfig'
        description: Global Agent configuration for this tenant (default for all sessions)
      api_key:
        description: API key
        type: string
      business:
        description: Business
        type: string
      context_config:
        allOf:
        - $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.ContextConfig'
        description: Global Context configuration for this tenant (default for all
          sessions)
      conversation_config:
        allOf:
        - $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.ConversationConfig'
        description: Global Conversation configuration for this tenant (default for
          normal mode sessions)
      created_at:
        description: Creation time
        type: string
      deleted_at:
        allOf:
        - $ref: '#/definitions/gorm.DeletedAt'
        description: Deletion time
      description:
        description: Description
        type: string
      id:
        description: ID
        type: integer
      name:
        description: Name
        type: string
      retriever_engines:
        allOf:
        - $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.RetrieverEngines'
        description: Retriever engines
      status:
        description: Status
        type: string
      storage_quota:
        description: Storage quota (Bytes), default is 10GB, including vector, original
          file, text, index, etc.
        type: integer
      storage_used:
        description: Storage used (Bytes)
        type: integer
      updated_at:
        description: Last updated time
        type: string
      web_search_config:
        allOf:
        - $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.WebSearchConfig'
        description: Global WebSearch configuration for this tenant
    type: object
...
  github_com_UniverseHappiness_WiseDx_internal_types.User:
    properties:
      avatar:
        description: Avatar URL of the user
        type: string
      can_access_all_tenants:
        description: Whether the user can access all tenants (cross-tenant access)
        type: boolean
      created_at:
        description: Creation time of the user
        type: string
      deleted_at:
        allOf:
        - $ref: '#/definitions/gorm.DeletedAt'
        description: Deletion time of the user
      email:
        description: Email address of the user
        type: string
      id:
        description: Unique identifier of the user
        type: string
      is_active:
        description: Whether the user is active
        type: boolean
      tenant:
        allOf:
        - $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.Tenant'
        description: Association relationship, not stored in the database
      tenant_id:
        description: Tenant ID that the user belongs to
        type: integer
      updated_at:
        description: Last updated time of the user
        type: string
      username:
        description: Username of the user
        type: string
    type: object
...
  github_com_UniverseHappiness_WiseDx_internal_types.WebSearchConfig:
    properties:
      api_key:
        description: API密钥（如果需要）
        type: string
      blacklist:
        description: 黑名单规则列表
        items:
          type: string
        type: array
      compression_method:
        description: 压缩方法：none, summary, extract, rag
        type: string
      document_fragments:
        description: 文档片段数量（用于RAG压缩）
        type: integer
      embedding_dimension:
        description: 嵌入维度（用于RAG压缩）
        type: integer
      embedding_model_id:
        description: RAG压缩相关配置
        type: string
      include_date:
        description: 是否包含日期
        type: boolean
      max_results:
        description: 最大搜索结果数
        type: integer
      provider:
        description: 搜索引擎提供商ID
        type: string
      rerank_model_id:
        description: 重排模型ID（用于RAG压缩）
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handler.EvaluationRequest:
    properties:
      chat_id:
        description: ID of chat model to use
        type: string
      dataset_id:
        description: ID of dataset to evaluate
        type: string
      knowledge_base_id:
        description: ID of knowledge base to use
        type: string
      rerank_id:
        description: ID of rerank model to use
        type: string
    type: object
...
info:
  contact:
    name: WeKnora Github
    url: https://github.com/UniverseHappiness/WiseDx
  description: WeKnora 知识库管理系统 API 文档
  termsOfService: http://swagger.io/terms/
  title: WeKnora API
  version: "1.0"
paths:
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: 修改当前用户的登录密码
      parameters:
      - description: 密码修改请求
        in: body
        name: request
        required: true
        schema:
          properties:
            new_password:
              type: string
            old_password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 修改密码
      tags:
      - 认证
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录并获取访问令牌
      parameters:
      - description: 登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.LoginResponse'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.AppError'
      summary: 用户登录
      tags:
      - 认证
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 撤销当前访问令牌并登出
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 用户登出
      tags:
      - 认证
  /auth/me:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 获取当前用户信息
      tags:
      - 认证
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌
      parameters:
      - description: 刷新令牌
        in: body
        name: request
        required: true
        schema:
          properties:
            refreshToken:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 新令牌
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 令牌无效
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.AppError'
      summary: 刷新令牌
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 注册新用户账号
      parameters:
      - description: 注册请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_types.RegisterResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.AppError'
        "403":
          description: 注册功能已禁用
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.AppError'
      summary: 用户注册
      tags:
      - 认证
  /auth/validate:
    get:
      consumes:
      - application/json
      description: 验证访问令牌是否有效
      produces:
      - application/json
      responses:
        "200":
          description: 令牌有效
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 令牌无效
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 验证令牌
      tags:
      - 认证
...
  /evaluation/:
    get:
      consumes:
      - application/json
      description: 根据任务ID获取评估结果
      parameters:
      - description: 评估任务ID
        in: query
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 评估结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 获取评估结果
      tags:
      - 评估
    post:
      consumes:
      - application/json
      description: 对知识库进行评估测试
      parameters:
      - description: 评估请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.EvaluationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 评估任务
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 执行评估
      tags:
      - 评估
...
securityDefinitions:
  ApiKeyAuth:
    description: 租户身份认证：输入 sk- 开头的 API Key
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: 用户登录认证：输入 Bearer {token} 格式的 JWT 令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"